Start working on addressing the Y2038 problem.
Plasma uses 32-bit timestamps in a few places, which will start causing
problems in a little over 12 years...  However, we don't want to break
backwards compatibility with the wire or file protocol, so this introduces
a new encoding which remains compatible until Jan 2038:

* For timestamps before max(int32), the same 32-bit encoding is used.
* For timestamps greater than max(int32), a dummy `0xFFFFFFFF` value is
  written, immediately followed by a 64-bit (signed) timestamp.  This is
  compatible with the definition of `time_t` on systems that use a 64-bit
  timestamp.

This means things will continue to work even with unpatched servers and
clients until the rollover point, which should hopefully give enough time
for everything to be updated and patched before crashing and burning.

This change specifically addresses the 32-bit timestamps everywhere
`plUnifiedTime` is used, and in the Vault nodes.  There are other 32-bit
timestamps left to convert (such as the server ping timers, and probably
others I'm forgetting about), but it's a start.