Remove hsThread::StartDetached, which is generally unsafe
After StartDetached is called, the hsThread can be destroyed while the
underlying thread is still running. This is unsafe in the general case,
because hsThread's thread proc captures the this pointer and calls
virtual methods (such as OnQuit) on the potentially-destroyed hsThread!

There is exactly one case where this method is safe, namely if the
hsThread is destroyed at the very end of its own OnQuit method. This was
done by pfPatcherWorker, the only hsThread implementation for which
StartDetached was used. pfPatcherWorker is now reimplemented using a
detached std::thread that owns all the data it needs, so the unsafe
hsThread::StartDetached method is no longer needed.